[PL+TLL] move initialization of plant module to group level
