+++ PopulationLib/Plant.py
@@ -54,23 +54,6 @@ class Plant:
             self.geometry["h_stem"] = initial_geometry["h_stem"]
         self.growth_concept_information = {}
 
-        self.iniPlantDynamicConcept()
-
-    def iniPlantDynamicConcept(self):
-        case = self.args.find("vegetation_model_type").text
-        if case == "Default":
-            from PlantModelLib.Default import Default as createGD
-        elif case == "Bettina":
-            from PlantModelLib.Bettina import Bettina as createGD
-        elif case == "Kiwi":
-            from PlantModelLib.Kiwi import Kiwi as createGD
-        elif case == "BettinaNetwork":
-            from PlantModelLib.BettinaNetwork import BettinaNetwork as createGD
-        else:
-            raise KeyError("Required plant dynamic concept not implemented.")
-        self.plant_dynamic_concept = createGD(self.args)
-        print(case + " plant dynamic concept initiated.")
-
     def getPosition(self):
         return self.x, self.y
